Fix Network/HTTP layer bugs: sizeof vs count, missing parens, uint8_t underflow, type mismatch, wrong HTTP reason phrase
- C-12: telnetServer.cpp - sizeof(TelnetCommands) gave byte size instead of element count
- H-17: NetworkHw.cpp - Ethernet accessor methods called without () yielded method pointers; MACAddress needs out-param
- H-18: telnetServer.cpp - trigger_ScanNodes declared bool but had no return statement
- H-19: httpServer.cpp - FindParameter: uint8_t underflow when param has no '=' separator
- H-20: httpServer.h - GetParameter declared with unsigned* but defined with uint16_t*
- M-10: httpServer.cpp - SendResponse424 sent "424 OK" instead of "424 Failed Dependency"

+++ Network/httpServer.h
@@ -26,7 +26,7 @@ class tHttpServlet
     void SendVersionAndPageClose();
 
     bool GetParameter(const char * Param) { return (NULL != FindParameter(Param,NULL,NULL)); }
-    bool GetParameter(const char * Param, unsigned *pValue);
+    bool GetParameter(const char * Param, uint16_t *pValue);
 
 
   private:
